feat: findAllBySearch - 키워드 검색 타입 별로 분별 / 기간 searchStartDate, searchEndDate null 값인 경우도 검색 가능하도록 변경

+++ src/main/java/org/example/expert/domain/todo/dto/response/TodoSearchResponse.java
@@ -1,15 +1,17 @@
 package org.example.expert.domain.todo.dto.response;
 
+import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 public class TodoSearchResponse {
 
     private final String title;
-    private final Integer managerCnt;
-    private final Integer commentCnt;
+    private final Long managerCnt;
+    private final Long commentCnt;
 
-    public TodoSearchResponse(String title, Integer managerCnt, Integer commentCnt) {
+    public TodoSearchResponse(String title, Long managerCnt, Long commentCnt) {
         this.title = title;
         this.managerCnt = managerCnt;
         this.commentCnt = commentCnt;
